# pup_patrol_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pup_patrol",
        "version": "1"
    },
    "paths": {
...
        "/api/informations": {
            "get": {
                "tags": [
                    "information"
                ],
                "summary": "informations 테이블",
                "description": "informations 테이블에 등록된 모든 데이터를 data라는 키를 가진 배열로 반환",
                "operationId": "3e80335d2b00defbcc436f0a8c98d8e2",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "information"
                ],
                "summary": "informations 테이블에 데이터 추가",
                "description": "informations 테이블에 데이터를 추가함",
                "operationId": "feb3c3058246ec9741b13229ab864342",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "informationDogName": {
                                        "description": "강아지 이름",
                                        "type": "string",
                                        "example": "골든 리트리버"
                                    },
                                    "informationDogCharacter": {
                                        "description": "강아지 성격",
                                        "type": "string",
                                        "example": "온화하고 영리함"
                                    },
                                    "informationMinSize": {
                                        "description": "평균 사이즈 범위의 시작 값",
                                        "type": "integer",
                                        "example": "56"
                                    },
                                    "informationMaxSize": {
                                        "description": "평균 사이즈 범위의 끝 값",
                                        "type": "integer",
                                        "example": "61"
                                    },
                                    "informationMinCost": {
                                        "description": "평균 분양가 범위의 시작 값",
                                        "type": "integer",
                                        "example": "25"
                                    },
                                    "informationMaxCost": {
                                        "description": "평균 분양가 범위의 끝 값 ",
                                        "type": "integer",
                                        "example": "80"
                                    },
                                    "informationDogText": {
                                        "description": "강아지 설명",
                                        "type": "string",
                                        "example": "골든 리트리버는 온화하며 특히 순종적이고 영리합니다. 사람에게 부드럽고 우호적입니다."
                                    },
                                    "informationDogGeneticillness": {
                                        "description": "유전병 정보",
                                        "type": "string",
                                        "example": "고관절 이형성증, 백내장, 암"
                                    },
                                    "informationCaution": {
                                        "description": "주의사항",
                                        "type": "string",
                                        "example": "사람에게는 착하지만, 다른 개들에게는 사나울 수 있습니다."
                                    },
                                    "informationImageName": {
                                        "description": "이미지 파일 이름",
                                        "type": "string",
                                        "example": "Golden Retriever.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "information이 성공적으로 추가됨"
                    }
                }
            }
        },
        "/api/informations/{information}": {
            "get": {
                "tags": [
                    "information"
                ],
                "summary": "파라미터로 넘어온 id를 가지는 특정 information 데이터 조회",
                "description": "파라미터로 넘어온 id를 가지는 특정 information 데이터를 data라는 키를 가진 객체로 반환",
                "operationId": "723e948e6e8caa9131225d15fe8bbdab",
                "parameters": [
                    {
                        "name": "information",
                        "in": "path",
                        "description": "information의 id",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "한 강아지의 information 데이터"
                    }
                }
            },
            "put": {
                "tags": [
                    "information"
                ],
                "summary": "information데이터 수정",
                "description": "한 개의 information데이터를 수정함 ",
                "operationId": "ffa6dbb2fd63839551ae39b0be320a43",
                "parameters": [
                    {
                        "name": "information",
                        "in": "path",
                        "description": "information의 아이디",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "informationDogName": {
                                        "description": "강아지 이름",
                                        "type": "string",
                                        "example": "골든 리트리버"
                                    },
                                    "informationDogCharacter": {
                                        "description": "강아지 성격",
                                        "type": "string",
                                        "example": "온화하고 영리함"
                                    },
                                    "informationMinSize": {
                                        "description": "평균 사이즈 범위의 시작 값",
                                        "type": "integer",
                                        "example": "56"
                                    },
                                    "informationMaxSize": {
                                        "description": "평균 사이즈 범위의 끝 값",
                                        "type": "integer",
                                        "example": "61"
                                    },
                                    "informationMinCost": {
                                        "description": "평균 분양가 범위의 시작 값",
                                        "type": "integer",
                                        "example": "25"
                                    },
                                    "informationMaxCost": {
                                        "description": "평균 분양가 범위의 끝 값 ",
                                        "type": "integer",
                                        "example": "80"
                                    },
                                    "informationDogText": {
                                        "description": "강아지 설명",
                                        "type": "string",
                                        "example": "골든 리트리버는 온화하며 특히 순종적이고 영리합니다. 사람에게 부드럽고 우호적입니다."
                                    },
                                    "informationDogGeneticillness": {
                                        "description": "유전병 정보",
                                        "type": "string",
                                        "example": "고관절 이형성증, 백내장, 암"
                                    },
                                    "informationCaution": {
                                        "description": "주의사항",
                                        "type": "string",
                                        "example": "사람에게는 착하지만, 다른 개들에게는 사나울 수 있습니다."
                                    },
                                    "informationImageName": {
                                        "description": "이미지 파일 이름",
                                        "type": "string",
                                        "example": "Golden Retriever.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "수정이 된 경우"
                    }
                }
            },
            "delete": {
                "tags": [
                    "information"
                ],
                "summary": "특정 information 데이터 삭제",
                "description": "하나의 information을 삭제함",
                "operationId": "c69464f4e9cb00098462eca5d5b88502",
                "parameters": [
                    {
                        "name": "information",
                        "in": "path",
                        "description": "강아지의 id",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "information 삭제에 성공한 경우"
                    }
                }
            }
        },
...
    },
...
}